requirements.txt
torch
torchvision
opencv-python
pillow
tqdm
numpy
scikit-learn